Last sync was 17d50f07 (2023-10-28); upstream was 468 commits ahead.

Upstream reads indirect objects held in object streams by their recorded
byte offsets. The fork read them sequentially, so a reference could resolve
to a neighbouring object's content: outline nodes appeared to point at /Link
annotations instead of actions, and destinations that did resolve landed on
the wrong page. That is the root cause behind WS2267, WS2271 and WS2468,
which patched the symptoms in PdfTokenScanner and ActionProvider.

Conflict resolution:
- Took upstream for PdfTokenScanner, PagesFactory and the cross reference
  subsystem (CrossReferenceParser and CrossReferenceObjectOffsetValidator
  were deleted upstream and replaced by FirstPassParser/XrefStreamParser).
- Kept the OnGettingBookmarkAction hook in PdfDocument, which the TOC
  extractor needs to read bookmark colour and style.
- Kept the OBDS tolerance customizations in ActionProvider (missing /S) and
  BookmarksProvider (title not required).
- Retired the WS2125 page 0 fallback: upstream models a destination-less
  outline node as a ContainerBookmarkNode, so callers ask for those with
  allowContainerNode instead of every caller getting a fake page 0 entry.

Target frameworks follow upstream (net47 is gone, net471 replaces it) and
the packages now require System.Memory 4.6.0 and Microsoft.Bcl.HashCode 6.0.0.
Version set to 0.1.9-alpha011.

Tests: 4383 passed, 3 failed - the pre-existing Standard14WritingFontTests
locale failures, unrelated to this merge.

+++ src/UglyToad.PdfPig.Tests/Writer/PdfPageBuilderTests.cs
@@ -1,19 +1,12 @@
 ﻿namespace UglyToad.PdfPig.Tests.Images
 {
-    using System;
-    using System.IO;
-    using System.Linq;
     using UglyToad.PdfPig;
-    using UglyToad.PdfPig.Content;
     using UglyToad.PdfPig.Core;
     using UglyToad.PdfPig.Fonts.Standard14Fonts;
     using UglyToad.PdfPig.Writer;
-    using Xunit;
-
+ 
     public class PdfPageBuilderTests
     {
-
-
         [Fact]
         public void CanAddPng()
         {
@@ -42,9 +35,9 @@
                     var dataPNG = LoadPng("4-16bitRGBA.png");
                     page4.AddPng(dataPNG, new PdfRectangle(0, 0, 595, 842));
                 }
+
                 pdfBytes = pdfBuilder.Build();
             }
-
 
             File.WriteAllBytes(@"PdfPageBuilderTests_CanAddPng.pdf", pdfBytes);
 
@@ -93,10 +86,6 @@
                     Assert.Equal(8, image1.BitsPerComponent);
                 }
             }
-            
-            
-
-
         }
 
 
@@ -245,9 +234,9 @@
 
             var page = pdfBuilder.AddPage(595d, 842d); // A4
             var dataPNG = LoadPng(imageFileName, subfolderName);
-            page.DrawRectangle(borderPlacement, (decimal)imagePlacement.Width, (decimal)imagePlacement.Height, 3, true);
+            page.DrawRectangle(borderPlacement, imagePlacement.Width, imagePlacement.Height, 3, true);
             page.AddPng(dataPNG, imagePlacement);
-            page.AddText(imageFileName, 12m, labelPlacement, font);
+            page.AddText(imageFileName, 12, labelPlacement, font);
         }
 
         private static byte[] LoadPng(string name, string subfolderName = null)
